Refactor CQL handling and fix identifier escaping
Move escape and dequote functions to CqlParsingRuleSet superclass for
better reusability and cleaner inheritance. Fix escape_name() to use
double-quote escaping for CQL identifiers and update tests accordingly.

patch by Jai Bheemsen Dhanwada; reviewed by [Reviewers] for CASSANDRA-20400
